api/requirements.txt
asgiref==3.9.1
cryptography==44.0.0
Django==5.2.5
django-admin-dracula==0.0.3
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3

# Performance and filtering
django-filter==24.3
django-cors-headers==4.6.0

# Monitoring and logging
python-json-logger==2.0.7

# Testing dependencies
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

# Code quality dependencies
black==24.10.0
flake8==7.1.1
isort==5.13.2
